refactor: 인증 집계 쿼리 반환 타입 불일치 수정 (Integer -> Long) #90
- Repository: JPQL COUNT 쿼리 스펙에 맞춰 반환 타입을 Long으로 변경
- Service: Long 타입을 받아 int로 변환하여 처리하도록 로직 수정

+++ src/main/java/com/hrr/backend/domain/verification/repository/VerificationRepository.java
@@ -65,7 +65,7 @@ public interface VerificationRepository extends JpaRepository<Verification, Long
             "WHERE r.round.id = :roundId " +
             "AND v.status = :status " +
             "AND v.createdAt BETWEEN :start AND :end")
-    Integer countDistinctCertifiers(
+    Long countDistinctCertifiers(
             @Param("roundId") Long roundId,
             @Param("status") VerificationStatus status,
             @Param("start") LocalDateTime start,
